move random submissionView f(X)s into utils
