+++ src/components/forms/SubmissionView.js
@@ -15,8 +15,12 @@ import { Alerts } from './../ui/Alerts';
 // We will cull util/file to only have some/minor fxns, and leave rest in 4DN repo.
 import { getLargeMD5 } from '../util/file';
 import { Detail } from './../ui/ItemDetailList';
+import { buildContext, findFieldFromContext, gatherLinkToTitlesFromContextEmbedded, modifyContextInPlace,
+    findParentFromHierarchy, flattenHierarchy, modifyHierarchy, searchHierarchy, trimHierarchy, replaceInHierarchy,
+    removeNulls, sortPropFields
+} from '../util/submission-view';
 
-import { SubmissionTree, fieldSchemaLinkToType, fieldSchemaLinkToPath } from './components/SubmissionTree';
+import { SubmissionTree, fieldSchemaLinkToType } from './components/SubmissionTree';
 import { BuildField, AliasInputField, isValueNull } from './components/submission-fields';
 
 /**
@@ -2352,500 +2356,4 @@ class RoundTwoDetailPanel extends React.PureComponent {
 
         );
     }
-}
-
-/***** MISC. FUNCIONS *****/
-
-
-
-
-/**
- * Build context based off an object's and populate values from
- * pre-existing context. Empty fields are given null value.
- * All linkTo fields are added to objList.
- * If initObjs provided (edit or clone functionality), pre-existing objs will be added.
- * Also checks user info to see if user is admin, which affects which fields are displayed.
- *
- * @returns {Object} A new object represent context.
- */
-export function buildContext(context, itemSchema, objList=null, edit=false, create=true, initObjs=null){
-    const built = {};
-    const userGroups = JWT.getUserGroups();
-    const fields = itemSchema.properties ? _.keys(itemSchema.properties) : [];
-
-    _.forEach(fields, function(field, i){
-        const fieldSchema = object.getNestedProperty(itemSchema, ['properties', field], true);
-
-        if (!fieldSchema){
-            return;
-        }
-
-        if (fieldSchema.exclude_from && ((Array.isArray(fieldSchema.exclude_from) && _.contains(fieldSchema.exclude_from,'FFedit-create')) || fieldSchema.exclude_from === 'FFedit-create')){
-            return;
-        }
-
-        // check to see if this field is a calculated prop
-        if (fieldSchema.calculatedProperty && fieldSchema.calculatedProperty === true){
-            return;
-        }
-
-        // check to see if permission == import items; if admin, allow import_items fields
-        if (fieldSchema.permission && fieldSchema.permission == "import_items"){
-            if (!_.contains(userGroups, 'admin')){
-                return;
-            }
-        }
-
-        // set value to context value if editing/cloning.
-        // if creating or value not present, set to null
-        if (edit){
-            if (context[field] === null || (fieldSchema.ff_flag && fieldSchema.ff_flag === "clear edit")){
-                built[field] = null;
-            } else {
-                built[field] = context[field] || null;
-            }
-        } else if (!create){ //clone
-            if (context[field] === null || (fieldSchema.ff_flag && fieldSchema.ff_flag === "clear clone")){
-                built[field] = null;
-            } else {
-                built[field] = context[field] || null;
-            }
-        } else {
-            built[field] = null;
-        }
-
-        if (objList !== null) {
-
-            let linkedProperty = fieldSchemaLinkToPath(fieldSchema); // Is it a linkTo (recursively or not)?
-            const roundTwoExclude = fieldSchema.ff_flag && fieldSchema.ff_flag == 'second round';
-
-            if ((linkedProperty !== null && typeof linkedProperty !== 'undefined') && !roundTwoExclude){ // If linkTo, add to our list, selecting a nice name for it first.
-                //var listTerm = fieldSchema.title ? fieldSchema.title : linked;
-                let fieldToStore = field;
-
-                linkedProperty = _.reject(linkedProperty, function(p){ return p === 'items' || p === 'properties'; });
-
-                if (linkedProperty.length > 0){
-                    fieldToStore += '.' + linkedProperty.join('.');
-                }
-
-                if (!_.contains(objList, fieldToStore)){
-                    objList.push(fieldToStore);
-                }
-
-                // add pre-existing linkTo objects
-                if (initObjs !== null && built[field] !== null){
-                    delvePreExistingObjects(initObjs, built[field], fieldSchema, fieldToStore);
-                }
-            }
-            objList.sort();
-        }
-    });
-
-    return built;
-}
-
-
-/**
- * Takes an initObjs array that it will fill with data for each existing
- * object in an edit/clone situation. json is json content for the field,
- * schema is the individual fields schema. Recursively handles objects and arrays
- */
-function delvePreExistingObjects(initObjs, json, fieldSchema, listTerm){
-    if (Array.isArray(json)){
-        for (var j=0; j < json.length; j++){
-            if (fieldSchema.items){
-                delvePreExistingObjects(initObjs, json[j], fieldSchema.items, listTerm);
-            }
-        }
-    } else if (json instanceof Object && json){
-        if(fieldSchema.properties){
-            _.keys(json).forEach(function(key, idx){
-                if(fieldSchema.properties[key]){
-                    delvePreExistingObjects(initObjs, json[key], fieldSchema.properties[key], listTerm);
-                }
-            });
-        }
-    } else if (_.contains(_.keys(fieldSchema),'linkTo')) { // non-array, non-object field. check schema to ensure there's a linkTo
-        initObjs.push({
-            'path'      : json,
-            'display'   : json,
-            'field'     : listTerm,
-            'type'      : fieldSchemaLinkToType(fieldSchema)
-        });
-    }
-}
-
-
-/** Sort a list of BuildFields first by required status, then by schema lookup order, then by title */
-function sortPropFields(fields){
-    var reqFields = [];
-    var optFields = [];
-
-    /** Compare by schema property 'lookup' meta-property, if available. */
-    function sortSchemaLookupFunc(a,b){
-        var aLookup = (a.props.schema && a.props.schema.lookup) || 750,
-            bLookup = (b.props.schema && b.props.schema.lookup) || 750,
-            res;
-
-        if (typeof aLookup === 'number' && typeof bLookup === 'number') {
-            //if (a.props.field === 'ch02_power_output' || b.props.field === 'ch02_power_output') console.log('X', aLookup - bLookup, a.props.field, b.props.field);
-            res = aLookup - bLookup;
-        }
-
-        if (res !== 0) return res;
-        else {
-            return sortTitle(a,b);
-        }
-    }
-
-    /** Compare by property title, alphabetically. */
-    function sortTitle(a,b){
-        if (typeof a.props.field === 'string' && typeof b.props.field === 'string'){
-            if(a.props.field.toLowerCase() < b.props.field.toLowerCase()) return -1;
-            if(a.props.field.toLowerCase() > b.props.field.toLowerCase()) return 1;
-        }
-        return 0;
-    }
-
-    _.forEach(fields, function(field){
-        if (!field) return;
-        if (field.props.required) {
-            reqFields.push(field);
-        } else {
-            optFields.push(field);
-        }
-    });
-
-    reqFields.sort(sortSchemaLookupFunc);
-    optFields.sort(sortSchemaLookupFunc);
-
-    return reqFields.concat(optFields);
-}
-
-
-
-
-function gatherLinkToTitlesFromContextEmbedded(context, idsToTitles = {}){
-    if (context['@id'] && context.display_title) {
-        if (typeof idsToTitles[context['@id']] !== 'undefined'){
-            // Seen already
-            return;
-        }
-        idsToTitles[context['@id']] = context.display_title;
-    }
-    _.values(context).forEach(function(value){
-        if (Array.isArray(value)){
-            value.forEach(function(arrItem){
-                if (!Array.isArray(arrItem) && arrItem && typeof arrItem === 'object'){
-                    gatherLinkToTitlesFromContextEmbedded(arrItem, idsToTitles);
-                }
-            });
-        } else if (value && typeof value === 'object'){
-            gatherLinkToTitlesFromContextEmbedded(value, idsToTitles);
-        }
-    });
-    return idsToTitles;
-}
-
-
-
-/**
- * Given the parent object key and a new object key, return a version
- * of this.state.keyHierarchy that includes the new parent-child relation.
- * Recursive function
- */
-var modifyHierarchy = function myself(hierarchy, keyIdx, parentKeyIdx){
-    _.keys(hierarchy).forEach(function(key, index){
-        if(key == parentKeyIdx){
-            hierarchy[parentKeyIdx][keyIdx] = {};
-        }else{
-            hierarchy[key] = myself(hierarchy[key], keyIdx, parentKeyIdx);
-        }
-    });
-    return hierarchy;
-};
-
-/** Remove given key from hierarchy. Recursive function. */
-var trimHierarchy = function myself(hierarchy, keyIdx){
-    if(hierarchy[keyIdx]){
-        delete hierarchy[keyIdx];
-    }else{
-        _.keys(hierarchy).forEach(function(key, index){
-            hierarchy[key] = myself(hierarchy[key], keyIdx);
-        });
-    }
-    return hierarchy;
-};
-
-/**
- * Returns the entire hierarchy below for the given keyIdx. keyIdx must be a
- * number (custom object). Recursive function.
- */
-var searchHierarchy = function myself(hierarchy, keyIdx){
-    if(!hierarchy) return null;
-    var found_hierarchy = null;
-    _.keys(hierarchy).forEach(function(key, index){
-        if(key == keyIdx){
-            found_hierarchy = hierarchy[key];
-        }else{
-            var test = myself(hierarchy[key], keyIdx);
-            if(test !== null){
-                found_hierarchy = test;
-            }
-        }
-    });
-    return found_hierarchy;
-};
-
-/** Finds the key of direct parent for a given key in a hierarchy */
-var findParentFromHierarchy = function myself(hierarchy, keyIdx){
-    if(isNaN(keyIdx) || !hierarchy) return null;
-    var found_parent = null;
-    _.keys(hierarchy).forEach(function(key, index){
-        if(keyIdx in hierarchy[key]){
-            found_parent = key;
-        }else{
-            var test = myself(hierarchy[key], keyIdx);
-            if(test !== null) found_parent = test;
-        }
-    });
-    return found_parent;
-};
-
-/**
- * Replace a key with a different key in the hierarchy
- */
-var replaceInHierarchy = function myself(hierarchy, existingValueToFind, newValue){
-    if (typeof existingValueToFind === 'number') existingValueToFind = existingValueToFind + '';
-    _.keys(hierarchy).forEach(function(key, index){
-        if(key === existingValueToFind){
-            var downstream = hierarchy[key];
-            hierarchy[newValue] = downstream;
-            delete hierarchy[key];
-        }else{
-            hierarchy[key] = myself(hierarchy[key], existingValueToFind, newValue);
-        }
-    });
-    return hierarchy;
-};
-
-/** Return a list of all keys contained within a given hierarchy */
-var flattenHierarchy = function myself(hierarchy){
-    var found_keys = [];
-    _.keys(hierarchy).forEach(function(key, index){
-        if(!isNaN(key)) key = parseInt(key);
-        var sub_keys = myself(hierarchy[key]);
-        found_keys = _.union(found_keys, sub_keys, [key]);
-    });
-    return found_keys;
-};
-
-
-/**
- * Remove any field with a null value from given json context.
- * also remove empty arrays and objects
- *
- * @param {Object} context - Object representing an Item, with properties & values.
- * @returns {Object} The same context which was passed in, minus null-y values.
- */
-function removeNulls(context){
-    _.keys(context).forEach(function(key, index){
-        if (isValueNull(context[key])){
-            delete context[key];
-        } else if (Array.isArray(context[key])){
-            context[key] = _.filter(context[key], function(v){ return !isValueNull(v); });
-            // Recurse for any objects
-            context[key] = _.map(context[key], function(v){
-                return (v && typeof v === 'object') ? removeNulls(v) : v;
-            });
-        } else if (context[key] instanceof Object) {
-            context[key] = removeNulls(context[key]);
-        }
-    });
-    return context;
-}
-/**
- * Takes in a deep COPY of keyContext and returns an edited version of that copy with the item in splitField
- * at the arrayIdx changed to value. (I THINK?! MAYBE FUCK IDK) GOD WHY
- *
- * @param {[string]} splitField An array containing the results of .split('.') on a field (E.g. experiment_sets.experiment => ["experiment_sets","experiment"])
- * @param {object} currContext  A DEEP COPY of keyContext
- * @param {array} arrayIdx      An array containing the index of the item to be modified
- * @param {string} fieldType    An array containing the type of object to be modified (E.g., "linked object", "existing object", etc.)
- * @param {string} value        A string containing the @id path of the item to be updated to
- *
- * Per the name, this function modifies IN PLACE and will cause super duper side effects if used on state.
- * Always pass in an object.deepClone'd COPY of the context you'd like to modify.
- *
- */
-function modifyContextInPlace(splitField, currContext, arrayIdx, fieldType, value){
-    console.log(`calling modifyContextInPlace with`, splitField, currContext, arrayIdx, fieldType, value);
-    const splitFieldLeaf = splitField[splitField.length-1];
-    let arrayIdxPointer = 0;
-    const contextCopy = currContext;
-    let pointer = contextCopy;
-    let prevValue = null;
-    for (let i=0; i < splitField.length - 1; i++){
-        // console.log(splitField[i]);
-        // console.log("pointer at start of loop", pointer);
-        if(pointer[splitField[i]]){
-            pointer = pointer[splitField[i]];
-        }else{
-            // console.log(pointer[splitField[i]]);
-            // console.log(pointer);
-            console.error('PROBLEM CREATING NEW CONTEXT WITH: ', fieldType, value);
-            return;
-        }
-        // console.log("pointer after updating with", splitField[i], " :", pointer);
-        if(Array.isArray(pointer)){
-            // console.log("pointer is array");
-            // console.log("before switch", pointer);
-            pointer = pointer[arrayIdx[arrayIdxPointer]];
-            // console.log("arrayIdx[arrayIdxPointer]", arrayIdx[arrayIdxPointer]);
-            // console.log(pointer);
-            arrayIdxPointer += 1;
-        }
-    }
-    // console.log("after for loop", pointer);
-    if (Array.isArray(pointer[splitFieldLeaf]) && fieldType !== 'array'){
-        // console.log("value we're trying to set is inside of an array");
-        // move pointer into array
-        pointer = pointer[splitFieldLeaf];
-        prevValue = pointer[arrayIdx[arrayIdxPointer]];
-        if (value === null){ // delete this array itemfieldType
-            pointer.splice(arrayIdx[arrayIdxPointer], 1);
-        } else {
-            pointer[arrayIdx[arrayIdxPointer]] = value;
-        }
-    } else { // value we're trying to set is not inside an array at this point
-        // console.log("value we're trying to set is NOT inside of an array");
-        prevValue = pointer[splitFieldLeaf];
-        pointer[splitFieldLeaf] = value;
-    }
-    return { currContext, prevValue };
-}
-
-
-/**
- * Traverses context to find the field name of the object at a specific keyIndex in context.
- *
- * @param {object} contextToSearch   Top level keyContext to search through
- * @param {string} rootType          The schema-formatted type of Item at the root of this context; principal object's type (E.g. "Experiment" or "Cohort")
- * @param {string} schemas           An object containing all schemas
- * @param {number} keyIndexToFind    The key index of the item to find
- * @param {array}  keyLinkToFind     An array representing the path to the item being searched for
- *
- * This might work if you pass in a subContext and make sure the rootType refers to the correct subContext's type, but not tested so can't be sure.
- *
- * @returns {object} { splitField: string[], arrayIdx: number[] }
- *          splitField represents the field name,
- *          arrayIdx contains the indices of any arrays searched in order to find the object during traversal.
- */
-function findFieldFromContext(contextToSearch, rootType, schemas, keyIndexToFind = 1, keyLinkToFind = []){
-    console.log("calling findFieldFromContext with: ", ...arguments);
-    // Issue: no way to figure out if new-linkto being created or not.
-
-    let splitField = null;
-    let arrayIdx = [];
-
-    /*
-    const keyHierarchy = {
-        0 : {}
-    };
-    */
-
-    // function isLinkTo(valString){
-    //     if (typeof valString !== "string") return false;
-    //     const matched = valString.match(/\/(.*?)\/(.*?)\//);
-    //     if (!matched) return false;
-    //     const [ , itemType, itemIdentifier ] = matched;
-    //     if (itemType && itemIdentifier) {
-    //         return itemType;
-    //     }
-    //     return false;
-    // }
-
-    /**
-     * Recursive function used to scrape through the context.
-     *
-     * @param {object} context          The keyContext object (or nested context object) to search
-     * @param {number} contextKey       The key in keyContext (or current nested context object) being searched
-     * @param {object} contextSchema    The schema for the type of object that is being searched
-     * @param {array}  currFieldParts   An array containing the previous contextKeys searched to get to this context
-     * @param {array}  arrIdx           Contains the indices of any arrays that were searched previously to get to this context,
-     *                                  in order from most to least recent
-     *
-     * When it finds an array or an object, it recursively searches for the field present until it finds it OR searches everything.
-     * Once the item being searched for (keyIndexToFind above) is found, updates findFieldFromContext's splitField and arrayIdx.
-     */
-    function scrapeFromContext(context, contextKey, contextSchema, currFieldParts = [], arrIdx = []){
-        console.log("calling scrapeFromcontext with", context, contextKey, contextSchema, currFieldParts, arrIdx);
-        splitField ? console.log("splitField is ", splitField) : null;
-        if (splitField) return; // recurses until it finds the field being sought
-
-        // Searches through the context passed in...
-        _.keys(context).forEach(function(propKey){
-
-            // Store the schema and value of the current nested context object being scraped
-            const propVal = context[propKey];
-            let propSchema = contextSchema[propKey];
-
-            if (Array.isArray(propVal)){ // If current item is an array of other items
-                propSchema = propSchema.items; // Make sure schema is synced as appropriate
-
-                // Traverse through each item in the array to see if field is present
-                propVal.forEach(function(propValItem, idxInArray){
-
-                    // If the items in this array are other linked objects, recurse and search them, too.
-                    if (propValItem !== null && typeof propValItem === "object"){
-                        console.log(`Found a new object. Scraping... ${contextKey}.${propKey}`);
-                        // NOTE: This breaks down when encountering nested arrays (more than 1 array deep).
-                        // But this occurs already other places so w.e. TODO: Fix this, if necessary
-                        scrapeFromContext(propValItem, propKey, propSchema.properties, [ ...currFieldParts, propKey ], [...arrIdx, idxInArray]);
-                        return;
-                    }
-
-                    // If the field is matching the item that is currently being sought, update splitField and arrayIdx
-                    if (keyIndexToFind === propValItem){
-                        const isCorrectLinkTo = keyLinkToFind.indexOf(propSchema.linkTo) > -1;
-                        if (isCorrectLinkTo) {
-                            splitField = [ ...currFieldParts, propKey ];
-                            arrayIdx = [ ...arrIdx, idxInArray];
-                            //keyHierarchy[contextKey] = keyHierarchy[contextKey] || {};
-                            //keyHierarchy[contextKey][propValItem] = itemType;
-                            //keyTypes[contextKey] = keyTypes[contextKey] || {};
-                            //keyTypes[]
-                        }
-                    }
-                });
-            } else if (propVal !== null && typeof propVal === "object"){
-                // Sub-embedded object. Recurse and search keys
-                scrapeFromContext(propVal, propKey, propSchema.properties, [ ...currFieldParts, propKey ], arrIdx);
-            } else {
-                // Assmuming this is a field; check to see if it matches
-                if (keyIndexToFind === propVal){
-                    const isCorrectLinkTo = keyLinkToFind.indexOf(propSchema.linkTo) > -1;
-                    if (isCorrectLinkTo) {
-                        splitField = [ ...currFieldParts, propKey ];
-                        arrayIdx = arrIdx;
-                        // keyHierarchy[contextKey] = keyHierarchy[contextKey] || {};
-                        // keyHierarchy[contextKey][propVal] = itemType;
-                    }
-                }
-            }
-        });
-    }
-
-    // _.keys(contextToSearch).forEach(function(propKey){
-        // Bleh need to traverse every sub object and array here
-        scrapeFromContext(contextToSearch, null, schemas[rootType].properties, []);
-    // });
-
-    console.log("returning splitfield: ", splitField);
-    console.log("returning arrayIdx: ", arrayIdx);
-    return {
-        splitField,
-        arrayIdx
-    };
 }
